@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,7 +30,7 @@ dependencies:
   bloc: ^8.0.3
   cupertino_icons: ^1.0.2
   equatable: ^2.0.3
-  firebase_core: ^1.17.1
+  firebase_core: ^2.13.1
   flutter:
     sdk: flutter
   flutter_bloc: ^8.0.1
@@ -38,6 +38,8 @@ dependencies:
 dev_dependencies:
   flutter_lints: ^2.0.0
   flutter_test:
+    sdk: flutter
+  integration_test:
     sdk: flutter
 
 # For information on the generic Dart part of this file, see the
